requirements.txt
requests==2.18.4
torchvision==0.4.2
numpy==1.17.3
Flask==1.1.1
gunicorn==19.9.0
https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
Pillow==6.2.2
flask_cors==3.0.8